Segment-Clothing/requirements.txt
transformers
torch
pillow
numpy